+++ client/src/components/Form.vue
@@ -5,7 +5,7 @@
     </BulmaQuickView>
     <div class="container" v-if="formIsReady">
       <div class="columns">
-        <div class="column">
+        <div class="column is-clipped">
           <h1 class="title">{{ currentForm.name }} <span v-if="currentForm.help" class="tag is-info is-clickable" @click="showHelp=!showHelp"><span class="icon"><font-awesome-icon icon="question-circle" /></span><span v-if="showHelp">Hide help</span><span v-else>Show help</span></span></h1>
           <article v-if="currentForm.help && showHelp" class="message is-info">
             <div class="message-body content"><VueShowdown :markdown="currentForm.help" flavor="github" :options="{ghCodeBlocks:true}" /></div>
@@ -16,6 +16,9 @@
             </span>
             <span>Show Extravars</span>
           </button>
+          <span v-if="isAdmin" class="icon is-clickable is-pulled-right" :class="{'has-text-success':!showHidden,'has-text-danger':showHidden}" @click="showHidden=!showHidden">
+              <font-awesome-icon :icon="(showHidden?'search-minus':'search-plus')" />
+          </span>
           <button @click="$emit('rerender')" class="button is-warning is-small mr-3">
             <span class="icon">
               <font-awesome-icon icon="redo" />
@@ -40,6 +43,14 @@
                       <label class="label has-text-primary">{{ field.label }} <span v-if="field.required" class="has-text-danger">*</span>
                         <span class="is-pulled-right">
                           <span
+                            class="icon is-clickable"
+                            :class="{'has-text-success':!fieldOptions[field.name].debug,'has-text-danger':fieldOptions[field.name].debug}"
+                            @click="setExpressionFieldDebug(field.name,!fieldOptions[field.name].debug)"
+                            v-if="field.expression && isAdmin"
+                          >
+                            <font-awesome-icon :icon="(fieldOptions[field.name].debug?'search-minus':'search-plus')" />
+                          </span>
+                          <span
                             class="icon is-clickable has-text-success"
                             @click="setExpressionFieldEditable(field.name,true)"
                             v-if="field.editable && field.type=='expression' && !fieldOptions[field.name].editable && !fieldOptions[field.name].viewable"
@@ -62,7 +73,7 @@
                           </span>
                           <span
                             class="icon is-clickable has-text-info"
-                            @click="clip($v.form[field.name].$model)"
+                            @click="clip((field.type=='expression')?$v.form[field.name].$model:queryresults[field.name])"
                             v-if="(field.type=='expression' || field.type=='query') && fieldOptions[field.name].viewable && !fieldOptions[field.name].editable"
                           >
                             <font-awesome-icon icon="copy" />
@@ -77,6 +88,24 @@
 
                         </span>
                       </label>
+                      <div class="mb-3"
+                        @dblclick="clip(field.expression,true)"
+                        v-if="field.expression && fieldOptions[field.name].debug">
+                        <highlight-code
+                          lang="javascript"
+                          :code="field.expression"
+                        />
+                      </div>
+
+                      <div class="mb-3"
+                        @dblclick="clip(fieldOptions[field.name].expressionEval,true)"
+                        v-if="field.expression && fieldOptions[field.name].debug && dynamicFieldStatus[field.name]!='fixed'">
+                        <highlight-code
+                          lang="javascript"
+                          :code="fieldOptions[field.name].expressionEval"
+                        />
+                      </div>
+
                       <!-- type = checkbox -->
                       <div v-if="field.type=='checkbox'">
                         <BulmaCheckRadio checktype="checkbox" v-model="$v.form[field.name].$model" :name="field.name" :type="{'is-danger is-block':$v.form[field.name].$invalid}" :label="field.placeholder" @change="evaluateDynamicFields(field.name)" />
@@ -104,9 +133,11 @@
                           :sticky="field.sticky||false"
                           >
                         </BulmaAdvancedSelect>
+
                         <div @dblclick="setExpressionFieldViewable(field.name,false)" v-if="fieldOptions[field.name].viewable" class="box limit-height mb-3">
                           <vue-json-pretty :data="queryresults[field.name]||[]"></vue-json-pretty>
                         </div>
+
                       </div>
                       <!-- type = radio -->
                       <div v-if="field.type=='radio'" >
@@ -129,9 +160,11 @@
                             <p @dblclick="setExpressionFieldViewable(field.name,true)" v-if="!fieldOptions[field.name].editable && !field.isHtml" class="input has-text-info" :class="{'is-danger':$v.form[field.name].$invalid}" v-text="stringify($v.form[field.name].$model)"></p>
                             <p @dblclick="setExpressionFieldViewable(field.name,true)" v-if="!fieldOptions[field.name].editable && field.isHtml" class="input has-text-info" :class="{'is-danger':$v.form[field.name].$invalid}" v-html="stringify($v.form[field.name].$model)"></p>
                           </div>
-                          <div @dblclick="setExpressionFieldViewable(field.name,false)" v-else class="box limit-height mb-3">
+
+                          <div @dblclick="setExpressionFieldViewable(field.name,false)" v-else  class="box limit-height mb-3">
                             <vue-json-pretty :data="$v.form[field.name].$model"></vue-json-pretty>
                           </div>
+
                         </div>
                         <!-- type = text or password-->
                         <input v-if="field.type=='text' || field.type=='password'"
@@ -239,7 +272,7 @@
             <span>Close output</span>
           </button>
         </div>
-        <div v-if="showJson" class="column">
+        <div v-if="showJson" class="column is-clipped">
           <h1 class="title">Extravars</h1>
           <button @click="showJson=false" class="button is-danger is-small">
             <span class="icon">
@@ -259,7 +292,7 @@
             </span>
             <span>Copy to clipboard</span>
           </button>
-          <div class="box mt-4">
+          <div class="box mt-4 is-limited">
             <vue-json-pretty :data="formdata"></vue-json-pretty>
           </div>
         </div>
@@ -277,6 +310,12 @@
   import BulmaQuickView from './BulmaQuickView.vue'
   import BulmaCheckRadio from './BulmaCheckRadio.vue'
   import BulmaEditTable from './BulmaEditTable.vue'
+
+  import 'highlight.js/styles/monokai-sublime.css'
+  import VueHighlightJS from 'vue-highlight.js';
+  import javascript from 'highlight.js/lib/languages/javascript';
+  import vue from 'vue-highlight.js/lib/languages/vue';
+
   import Helpers from './../lib/Helpers'
   import Copy from 'copy-to-clipboard'
   import 'vue-json-pretty/lib/styles.css';
@@ -285,13 +324,20 @@
 
   Vue.use(Vuelidate)
   Vue.use(VueShowdown)
+  Vue.use(VueHighlightJS, {
+    // Register only languages that you want
+      languages: {
+        javascript
+      }
+    });
 
   export default{
     name:"Form",
     components:{VueJsonPretty,BulmaAdvancedSelect,BulmaEditTable,BulmaCheckRadio,BulmaQuickView},
     props:{
       currentForm:{type:Object},
-      constants:{type:Object}
+      constants:{type:Object},
+      isAdmin:{type:Boolean}
     },
     data(){
       return  {
@@ -324,6 +370,7 @@
           validationsLoaded:false,
           timeout:undefined,     // determines how long we should show the result of run
           showHelp:false,
+          showHidden:false,
           ansibleJobId:undefined,
           awxJobId:undefined,
           abortTriggered:false
@@ -440,6 +487,9 @@
       setExpressionFieldViewable(fieldname,value){
         Vue.set(this.fieldOptions[fieldname],'viewable',value)   // flag editable
       },
+      setExpressionFieldDebug(fieldname,value){
+        Vue.set(this.fieldOptions[fieldname],'debug',value)   // flag editable
+      },
       stringify(v){
         if(v){
           if(Array.isArray(v)){
@@ -453,9 +503,14 @@
           return v
         }
       },
-      clip(v){
+      clip(v,doNotStringify=false){
         try{
-          Copy(JSON.stringify(v))
+          if(doNotStringify){
+            Copy(v)
+          }else{
+            Copy(JSON.stringify(v))
+          }
+
           this.$toast.success("Copied to clipboard")
         }catch(e){
           this.$toast.error("Error copying to clipboard : \n" + e)
@@ -463,11 +518,12 @@
 
       },
       filterfieldsByGroup(group){                   // creates a list of field per group
+        var ref=this
         return this.currentForm.fields.filter(function (el) {
           return (
             (("group" in el && el.group === group)
               || !("group" in el) && (group==""))
-            && el.hide!==true)
+            && (el.hide!==true || ref.showHidden))
         });
       },
       checkDependencies(field){
@@ -788,6 +844,7 @@
                   // set flag running
                   ref.setFieldStatus(item.name,"running",false)
                   placeholderCheck = ref.replacePlaceholders(item)     // check and replace placeholders
+                  Vue.set(ref.fieldOptions[item.name],"expressionEval",placeholderCheck.value||"undefined")
                   // console.log(`[${item.name}] 1 : ${placeholderCheck.value}`)
                   if(placeholderCheck.value!=undefined){                       // expression is clean ?
                       // console.log(`[${item.name}] 2 : ${placeholderCheck.value}`)
@@ -1347,7 +1404,7 @@
       }
       this.$v.form.$reset();
       // find all variable dependencies
-      this.findVariableDependencies()      
+      this.findVariableDependencies()
       this.findVariableDependentOf()
       this.startDynamicFieldsLoop();
     },
@@ -1363,6 +1420,14 @@ pre{
   white-space: -pre-wrap;      /* Opera 4-6 */
   white-space: -o-pre-wrap;    /* Opera 7 */
   word-wrap: break-word;       /* Internet Explorer 5.5+ */
+}
+.is-limited {
+  text-overflow: ellipsis;
+  width:100%;
+  overflow: hidden;
+}
+.is-nowrap{
+  white-space:nowrap;
 }
 .limit-height{
   max-height:300px;
@@ -1390,9 +1455,6 @@ pre{
 @keyframes bounce-in {
   0% {
     transform: scale(0);
-  }
-  50% {
-    transform: scale(1.5);
   }
   100% {
     transform: scale(1);